Update view.md with new APIs for issue #929 (#931)
* Update view.md with new APIs for issue #929

☂️☂️☂️☂️☂️☂️ #929  ☂️☂️☂️☂️☂️☂️☂️ 

New APIs added:
-nextFocusDown
-nextFocusForward
-nextFocusLeft
-nextFocusUp
-clickable

I don't have much details on any of the `nextFocus` apis, the descriptions in the code are very vague and tell me to refer to the docs which is pretty funny because I'm adding them to the docs.


Is there any easier way to figure out what new stuff has been added? I pretty much found the new APIs in the middle of View.js file, it didn't seem to have any particular order. I was hoping that new Apis would at least be at the very bottom from oldest to newest going from top to bottom.

* alphabetize API list

* Better descriptions for nextFocus

* Update view.md

* Update view.md